requirements.txt
asgiref==3.8.1
Django==5.0.7
django-jazzmin==3.0.0
django-modeltranslation==0.19.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
setuptools==75.1.0
sqlparse==0.5.1
uritemplate==4.1.1
